[Scene]: 'cartesian-plane' scene is moved to React 'CartesianPlaneScene' component.
Changelog:
- old features is restored in new 'scene' component;
- scene classes moved to 'classes' folder;
- added some imports to classes;
- old 'scene.js' file is deleted now.
