+++ src/js/scenes/cartesian-plane/classes/graph.class.js
