+++ src/js/scenes/cartesian-plane/classes/primitive.class.js
